\name{gamboostLSS-package}
\alias{gamboostLSS-package}

\docType{package}
\title{
  Boosting Methods for GAMLSS models
}
\description{
  Boosting models for fitting generalized additive models for
  location, shape and scale (gamLSS models) to potentially high dimensional
  data.
}
\details{
\tabular{ll}{
Package: \tab gamboostLSS\cr
Version: \tab 0.5-0\cr
Date: \tab 2010-07-02\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Benjamin Hofner, Andreas Mayr, Nora Fenske, Matthias Schmid

Maintainer:  Benjamin Hofner <benjamin.hofner@imbe.med.uni-erlangen.de>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{ ...
}
\examples{
 ## ADD SOMETHING
}
