\name{gamboostLSS_intern}
\alias{gamboostLSS_intern}
\title{
  Call internal functions.
}
\description{
  Call one of the internal gamboostLSS functions.
}
\usage{
gamboostLSS_intern(..., fun = c("check", "do_trace"))
}
\arguments{
  \item{\dots}{ Arguments to \code{fun}. }
  \item{fun}{ The name on an internal gamboostLSS function. }
}
\details{
  This function must not be called under any circumstances.
}
\keyword{internal}
